+++ src/main/java/com/abc/account/Account.java
@@ -2,7 +2,7 @@ package com.abc.account;
 
 import java.util.Collection;
 
-import com.abc.Transaction;
+import com.abc.transaction.Transaction;
 
 public interface Account {
