+++ src/main/java/com/abc/transaction/AbstractTransaction.java
@@ -1,4 +1,4 @@
-package com.abc;
+package com.abc.transaction;
 
 import java.util.Date;
