+++ src/test/java/com/abc/ReportTest.java
@@ -2,7 +2,6 @@ package com.abc;
 
 import com.abc.account.CheckingAccount;
 import com.abc.account.SavingsAccount;
-import com.abc.transaction.Bank;
 import org.junit.Before;
 import org.junit.Test;
